+++ src/modules/Majors/MajorDetails/MajorDetails.tsx
@@ -1,15 +1,83 @@
+import { Typography } from '@mui/material';
+import { styled } from '@mui/material/styles';
+import { DetailsTile, Divider } from '@UG/libs/components';
 import { Major } from '@UG/libs/types';
-import { useMemo } from 'react';
-import { usePrepareMajorDetails } from '../hooks';
+import { useEffect } from 'react';
+import { useSelector } from 'react-redux';
+import { useNavigate } from 'react-router-dom';
+import { StateType } from 'src/store';
+import { MajorDetailsInfo } from './MajorDetailsInfo';
+import { useGetMajors } from '../hooks';
 
-interface MajorDetailsProps {
-  major: Major;
+interface StateProps {
+  isLoading: boolean;
+  errorMessage: string | null;
+  majorDetails: Major | null;
+  majorsList: Major[];
 }
 
-export const MajorDetails = ({ major }: MajorDetailsProps) => {
-  const { prepareMajorDetails } = usePrepareMajorDetails();
+interface MajorDetailsProps {
+  id: string;
+}
 
-  const majorDetails = useMemo(() => prepareMajorDetails(major.content), [major.content, prepareMajorDetails]);
+const StyledTitle = styled(Typography)`
+  font-family: Montserrat;
+  font-size: 36px;
+  font-weight: 700;
+  line-height: 44px;
+  letter-spacing: 0em;
+  margin-bottom: 15px;
+  color: ${({ theme }) => theme.palette.secondary.dark};
+`;
 
-  return <>{majorDetails}</>;
+export const MajorDetails = ({ id }: MajorDetailsProps) => {
+  const { getMajorDetails } = useGetMajors();
+  const navigate = useNavigate();
+  const { isLoading, errorMessage, majorDetails } = useSelector<StateType, StateProps>(state => ({
+    isLoading: state.majors.isLoading,
+    errorMessage: state.majors.error,
+    majorDetails: state.majors.majorDetails,
+    majorsList: state.majors.majorsList,
+  }));
+
+  useEffect(() => {
+    getMajorDetails(id);
+  }, [id, getMajorDetails, navigate]);
+
+  //TODO change layout as soon as we get designs
+  if (!isLoading && errorMessage) {
+    return (
+      <>
+        <p style={{ marginTop: '150px' }}>{errorMessage}</p>
+      </>
+    );
+  }
+
+  //TODO change layout as soon as we get designs
+  if (isLoading && !errorMessage) {
+    return (
+      <>
+        <p style={{ marginTop: '150px' }}>loading...</p>
+      </>
+    );
+  }
+
+  //TODO change layout as soon as we get designs
+  if (!majorDetails) {
+    return (
+      <>
+        <p style={{ marginTop: '150px' }}>sorry but we couldn&apos;t find {id}</p>
+      </>
+    );
+  }
+
+  return (
+    <>
+      <StyledTitle data-cy="title">{majorDetails.name}</StyledTitle>
+      <Divider />
+      <DetailsTile>
+        <MajorDetailsInfo major={majorDetails} />
+      </DetailsTile>
+    </>
+  );
 };
